chore(redemptions): move verify redemptions actions and their tests
 - delete verifyRedemptionActions.js
 - add verify redemption actions in redeemPointsActions.js
 - delete verifyRedemptionActions.tests.js
 - add verifyRedemptionActions tests in redeemPointsActions.test.js

[Maintains #159326777]
