+++ tests/actions/redeemPointsActions.test.js
@@ -13,6 +13,7 @@ import {
   updateRedemptionFailure,
   updateRedemptionSuccess,
   updateRedemption,
+  verifyRedemption,
 } from '../../src/actions/redeemPointsAction';
 
 // types
@@ -26,6 +27,9 @@ import {
   UPDATE_REDEMPTION_REQUEST,
   UPDATE_REDEMPTION_FAILURE,
   UPDATE_REDEMPTION_SUCCESS,
+  VERIFY_REDEMPTION_SUCCESS,
+  VERIFY_REDEMPTION_FAILURE,
+  VERIFY_REDEMPTION_REQUEST,
 } from '../../src/types';
 
 // helpers
@@ -45,11 +49,11 @@ describe('Redeem Points Actions', () => {
   });
   afterEach(() => moxios.uninstall());
 
-  it('should dispatch redeem points request', () => {
+  it('should dispatch create redeem points request', () => {
     expect(createRedeemPointsRequest()).toEqual({ type: CREATE_REDEEM_POINTS_REQUEST });
   });
 
-  it('should dispatch redeem points failure', () => {
+  it('should dispatch create redeem points failure', () => {
     expect(createRedeemPointsFailure()).toEqual({ type: CREATE_REDEEM_POINTS_FAILURE });
   });
 
@@ -178,4 +182,55 @@ describe('Redeem Points Actions', () => {
     return store.dispatch(updateRedemption(updateData))
       .then(() => (expect(store.getActions()[1]).toEqual(expectedFailureAction)));
   });
+
+  describe('Verify redemption actions', () => {
+    it('dispatches VERIFY_REDEMPTION_SUCCESS after successfuly updating the redemption status to rejected', () => {
+      const updatedRedemption = {
+        ...redemption,
+        status: 'rejected'
+      }
+      const expectedActions = [
+        {
+          type: VERIFY_REDEMPTION_REQUEST,
+        },
+        {
+          type: VERIFY_REDEMPTION_SUCCESS,
+          redemption: updatedRedemption,
+        },
+      ];
+  
+      moxios.stubRequest(`${config.API_BASE_URL}/societies/redeem/verify/${updatedRedemption.id}`, {
+        status: 200,
+        response: { data: updatedRedemption },
+      });
+  
+      return store.dispatch(verifyRedemption(redemption.id, 'rejected')).then(() => {
+        expect(store.getActions()).toEqual(expectedActions);
+      });
+    });
+  
+    it('dispatches VERIFY_REDEMPTION_FAILURE when rejecting a redemption fails', () => {
+      const updatedRedemption = {
+        ...redemption,
+        status: 'rejected'
+      }
+      const expectedActions = [
+        {
+          type: VERIFY_REDEMPTION_REQUEST,
+        },
+        {
+          type: VERIFY_REDEMPTION_FAILURE,
+          error: new Error('Request failed with status code 401'),
+        },
+      ];
+  
+      moxios.stubRequest(`${config.API_BASE_URL}/societies/redeem/verify/${updatedRedemption.id}`, {
+        status: 401,
+      });
+  
+      return store.dispatch(verifyRedemption(redemption.id, 'rejected')).then(() => {
+        expect(store.getActions()).toEqual(expectedActions);
+      });
+    });
+  })
 });
